Update kaprekarUtils.c
changed int to double

+++ c_version/kaprekarUtils.c
@@ -33,7 +33,7 @@ int isKaprekar(int n) {
   long square = n * n;
 
   //compute the number of digits of the square
-  int numDigits =log10(square) + 1;
+  double numDigits =log10(square) + 1;
   long modulus = 1;
   long first, second;
